// src/app/interface/data-list.interface.ts
export interface User{
    id?: string
    user: string
    firstName: string
    middleName: string
    firstSurname: string
    secondLastName: string
    department?: string
    post?: string
    email: string


}

export interface Department{
  nameDepa: string
}

export interface Post{
  namePost: string
}